Update README.md with additional .zshrc mention
As of 2019, the default shell for MacOS is zsh . The README specifies you add the PATH to your .bash_profile, however this does not work if your default shell is zsh. To avoid having to activate the bash shell each time you run a cube_conversion, it is simpler to add to the PATH of the default shell, which is now zsh.